move json linter to Editor.service
