+++ web/src/services/Editor.service.ts
@@ -1,8 +1,9 @@
-import {SyntaxNode} from '@lezer/common/dist';
-import {EditorState} from '@codemirror/state';
+import {Diagnostic} from '@codemirror/lint';
+import {EditorState, Text} from '@codemirror/state';
 import {syntaxTree} from '@codemirror/language';
 import {EditorView} from '@codemirror/view';
 import {Completion, CompletionContext, CompletionResult} from '@codemirror/autocomplete';
+import {SyntaxNode} from '@lezer/common/dist';
 import {
   completeSourceAfter,
   operatorList,
@@ -32,6 +33,18 @@ interface IAutoCompleteProps {
   envEntryList?: IKeyValue[];
   customValueList?: string[];
   onSelect?(option: Completion): void;
+}
+
+const CUSTOM_SYNTAX_ERRORS_WHITE_LIST = [/Unexpected token \'\$\'/, /\${env:([\w-]+)}/];
+
+function getJsonErrorPosition(error: SyntaxError, doc: Text): number {
+  let m;
+  // eslint-disable-next-line no-cond-assign
+  if ((m = error.message.match(/at position (\d+)/))) return Math.min(Number(m[1]), doc.length);
+  // eslint-disable-next-line no-cond-assign
+  if ((m = error.message.match(/at line (\d+) column (\d+)/)))
+    return Math.min(doc.line(Number(m[1])).from + Number(m[2]) - 1, doc.length);
+  return 0;
 }
 
 const EditorService = () => ({
@@ -178,6 +191,34 @@ const EditorService = () => ({
     } catch (e) {
       return false;
     }
+  },
+
+  /// Calls
+  /// [`JSON.parse`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/parse)
+  /// on the document and, if that throws an error, reports it as a
+  /// single diagnostic.
+  jsonParseLinter() {
+    return (view: EditorView): Diagnostic[] => {
+      try {
+        JSON.parse(view.state.doc.toString());
+      } catch (e) {
+        if (!(e instanceof SyntaxError)) throw e;
+
+        const whiteListTest = CUSTOM_SYNTAX_ERRORS_WHITE_LIST.map(regex => regex.test(e.message));
+        if (whiteListTest.every(t => t)) return [];
+
+        const pos = getJsonErrorPosition(e, view.state.doc);
+        return [
+          {
+            from: pos,
+            message: e.message,
+            severity: 'error',
+            to: pos,
+          },
+        ];
+      }
+      return [];
+    };
   },
 });
